src/requirements.txt
lxml
pandas
sqlalchemy-aurora-data-api
